apps/worker/requirements.txt
python-dotenv
supabase
requests
opencv-python-headless
pytesseract
pytesseract-ocr
pytesseract